requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-extensions==3.1.1
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
numpy==1.20.1
pandas==1.2.1
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
